Conway's Game of Life: Restructure
- Rename cell.py to agents.py
- Flatten structure
- Remove unneeded viz
