+++ examples/basic/conways_game_of_life/Readme.md
@@ -19,19 +19,17 @@ Then open your browser to [http://127.0.0.1:8521/](http://127.0.0.1:8521/) and p
 
 ## Files
 
-* ``conways_game_of_life/cell.py``: Defines the behavior of an individual cell, which can be in two states: DEAD or ALIVE.
-* ``conways_game_of_life/model.py``: Defines the model itself, initialized with a random configuration of alive and dead cells.
-* ``conways_game_of_life/portrayal.py``: Describes for the front end how to render a cell.
-* ``conways_game_of_life/server.py``: Defines an interactive visualization.
-* ``run.py``: Launches the visualization
+* ``agents.py``: Defines the behavior of an individual cell, which can be in two states: DEAD or ALIVE.
+* ``model.py``: Defines the model itself, initialized with a random configuration of alive and dead cells.
+* ``portrayal.py``: Describes for the front end how to render a cell.
+* ``st_app.py``: Defines an interactive visualization using Streamlit.
 
 ## Optional
 
-*  ``conways_game_of_life/app.py``: can be used to run the simulation via the streamlit interface.
+*  ``conways_game_of_life/st_app.py``: can be used to run the simulation via the streamlit interface.
 * For this some additional packages like ``streamlit`` and ``altair`` needs to be installed.
-* Once installed, the app can be opened in the browser using : ``streamlit run app.py``
+* Once installed, the app can be opened in the browser using : ``streamlit run st_app.py``
 
 
 ## Further Reading
 [Conway's Game of Life](https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life)
-
