+++ examples/basic/conways_game_of_life/portrayal.py
