--- examples/basic/conways_game_of_life/run.py
@@ -1,3 +1,0 @@
-from conways_game_of_life.server import server
-
-server.launch(open_browser=True)
