+++ examples/basic/conways_game_of_life/server.py
@@ -1,7 +1,7 @@
 import mesa
 
-from .model import ConwaysGameOfLife
-from .portrayal import portrayCell
+from model import ConwaysGameOfLife
+from portrayal import portrayCell
 
 # Make a world that is 50x50, on a 250x250 display.
 canvas_element = mesa.visualization.CanvasGrid(portrayCell, 50, 50, 250, 250)
